FaceRecog/facerecog/requirements.txt
tensorflow
numpy
opencv
matplotlib
scikit-learn
scikit-image
Pillow